Avance 17

+++ src/modelo/PruebaLogicaNomina.java
@@ -17,27 +17,27 @@ public class PruebaLogicaNomina {
     }
 
     //TotalDevengado: Total a pagar cada quincena (sin salud ni pension)
-    public int totalD(int salarioBase, int diasTrabajados) {
+    public double totalD(double salarioBase, double diasTrabajados) {
 
-        int[] vectorV = validacion(salarioBase, diasTrabajados);
+        double[] vectorV = validacion(salarioBase, diasTrabajados);
 
-        int salarioD = vectorV[0];
-        int auxD = vectorV[1];
-        int totalDevengado = salarioD + auxD;
+        double salarioD = vectorV[0];
+        double auxD = vectorV[1];
+        double totalDevengado = salarioD + auxD;
 
         return totalDevengado;
     }
 
-    public double totalPagar(int salarioBase, int diasTrabajados) {
+    public double totalPagar(double salarioBase, double diasTrabajados) {
 
-        int[] vector = validacion(salarioBase, diasTrabajados);
-        int salario = vector[0];
+        double[] vector = validacion(salarioBase, diasTrabajados);
+        double salario = vector[0];
         double totalPagar = salario - (salario * 0.08);
         return totalPagar;
     }
 
     //Cesantias: Dado anualmente a un fondo de cesantias cada 14 de febrero
-    public double cesantias(int salarioBase, int diasTrabajados) {
+    public double cesantias(double salarioBase, double diasTrabajados) {
 
         //Entradas salario Base, auxilio de Transporte
         int auxTransp = 117172;
@@ -51,7 +51,7 @@ public class PruebaLogicaNomina {
     }
 
     //Prima: Dado al empleado cada 6 meses (Junio y Diciembre) en quincenas
-    public double prima(int diasTrabajados, int salarioBase) {
+    public double prima(double diasTrabajados, double salarioBase) {
 
         //Prima = (1/2 salarioBase/180)*diasTrabajados
         double prima = ((salarioBase / 2) / 180) * diasTrabajados;
@@ -60,7 +60,7 @@ public class PruebaLogicaNomina {
 
     //Vacaciones: Se pagan cuando el trabajador cumple 1 año en la empresa y
     //el pago corresponde a 15 dias de trabajo (salarioBase/2)
-    public double vacaciones(int salarioBase, int diasTrabajados) {
+    public double vacaciones(double salarioBase, double diasTrabajados) {
 
         double vacaciones;
         if (diasTrabajados == 360) {
@@ -74,7 +74,7 @@ public class PruebaLogicaNomina {
 
     //Intereses a Cesantias: Se pagan al trabajador anualmente sobre el
     //12% totales de las censantias cada 31 de Enero
-    public double interesCesantias(int salarioBase, int diasTrabajados) {
+    public double interesCesantias(double salarioBase, double diasTrabajados) {
 
         double cesantias = cesantias(salarioBase, diasTrabajados);
         double intCesantias = cesantias * 0.12;
@@ -82,10 +82,10 @@ public class PruebaLogicaNomina {
         return intCesantias;
     }
 
-    public int[] validacion(int salarioBase, int diasTrabajados) {
+    public double[] validacion(double salarioBase, double diasTrabajados) {
 
-        int salarioDevengado = 0;
-        int auxTranspDevengado = 0;
+        double salarioDevengado = 0;
+        double auxTranspDevengado = 0;
 
         if (diasTrabajados == 13) {
             salarioDevengado = salarioBase / 2;
@@ -93,7 +93,7 @@ public class PruebaLogicaNomina {
                 auxTranspDevengado = 117172 / 2;
             }
         }
-        int[] totalDevengado = {salarioDevengado, auxTranspDevengado};
+        double[] totalDevengado = {salarioDevengado, auxTranspDevengado};
 
         return totalDevengado;
     }
